requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
anyio==3.6.2
    # via jupyter-server
argon2-cffi==21.3.0
    # via
    #   jupyter-server
    #   nbclassic
    #   notebook
argon2-cffi-bindings==21.2.0
    # via argon2-cffi
arrow==1.2.3
    # via isoduration
asttokens==2.2.1
    # via stack-data
attrs==23.1.0
    # via jsonschema
backcall==0.2.0
    # via ipython
beautifulsoup4==4.12.2
    # via nbconvert
bleach==6.0.0
    # via nbconvert
cffi==1.15.1
    # via argon2-cffi-bindings
comm==0.1.3
    # via ipykernel
debugpy==1.6.7
    # via ipykernel
decorator==5.1.1
    # via ipython
defusedxml==0.7.1
    # via nbconvert
executing==1.2.0
    # via stack-data
fastjsonschema==2.16.3
    # via nbformat
fqdn==1.5.1
    # via jsonschema
idna==3.4
    # via
    #   anyio
    #   jsonschema
ipykernel==6.22.0
    # via
    #   nbclassic
    #   notebook
ipython==8.13.1
    # via ipykernel
ipython-genutils==0.2.0
    # via
    #   nbclassic
    #   notebook
isoduration==20.11.0
    # via jsonschema
jedi==0.18.2
    # via ipython
jinja2==3.1.2
    # via
    #   jupyter-server
    #   nbclassic
    #   nbconvert
    #   notebook
jsonpointer==2.3
    # via jsonschema
jsonschema[format-nongpl]==4.17.3
    # via
    #   jupyter-events
    #   nbformat
jupyter-client==8.2.0
    # via
    #   ipykernel
    #   jupyter-server
    #   nbclassic
    #   nbclient
    #   notebook
jupyter-core==5.3.0
    # via
    #   ipykernel
    #   jupyter-client
    #   jupyter-server
    #   nbclassic
    #   nbclient
    #   nbconvert
    #   nbformat
    #   notebook
jupyter-events==0.6.3
    # via jupyter-server
jupyter-server==2.5.0
    # via
    #   nbclassic
    #   notebook-shim
jupyter-server-terminals==0.4.4
    # via jupyter-server
jupyterlab-pygments==0.2.2
    # via nbconvert
markupsafe==2.1.2
    # via
    #   jinja2
    #   nbconvert
matplotlib-inline==0.1.6
    # via
    #   ipykernel
    #   ipython
mistune==2.0.5
    # via nbconvert
nbclassic==0.5.6
    # via notebook
nbclient==0.7.4
    # via nbconvert
nbconvert==7.3.1
    # via
    #   jupyter-server
    #   nbclassic
    #   notebook
nbformat==5.8.0
    # via
    #   jupyter-server
    #   nbclassic
    #   nbclient
    #   nbconvert
    #   notebook
nest-asyncio==1.5.6
    # via
    #   ipykernel
    #   nbclassic
    #   notebook
notebook==6.5.4
    # via -r requirements.in
notebook-shim==0.2.3
    # via nbclassic
packaging==23.1
    # via
    #   ipykernel
    #   jupyter-server
    #   nbconvert
pandocfilters==1.5.0
    # via nbconvert
parso==0.8.3
    # via jedi
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
platformdirs==3.5.0
    # via jupyter-core
prometheus-client==0.16.0
    # via
    #   jupyter-server
    #   nbclassic
    #   notebook
prompt-toolkit==3.0.38
    # via ipython
psutil==5.9.5
    # via ipykernel
ptyprocess==0.7.0
    # via
    #   pexpect
    #   terminado
pure-eval==0.2.2
    # via stack-data
py4j==0.10.9.7
    # via pyspark
pycparser==2.21
    # via cffi
pygments==2.15.1
    # via
    #   ipython
    #   nbconvert
pyrsistent==0.19.3
    # via jsonschema
pyspark==3.4.0
    # via -r requirements.in
python-dateutil==2.8.2
    # via
    #   arrow
    #   jupyter-client
python-json-logger==2.0.7
    # via jupyter-events
pyyaml==6.0
    # via jupyter-events
pyzmq==25.0.2
    # via
    #   ipykernel
    #   jupyter-client
    #   jupyter-server
    #   nbclassic
    #   notebook
rfc3339-validator==0.1.4
    # via
    #   jsonschema
    #   jupyter-events
rfc3986-validator==0.1.1
    # via
    #   jsonschema
    #   jupyter-events
send2trash==1.8.2
    # via
    #   jupyter-server
    #   nbclassic
    #   notebook
six==1.16.0
    # via
    #   asttokens
    #   bleach
    #   python-dateutil
    #   rfc3339-validator
sniffio==1.3.0
    # via anyio
soupsieve==2.4.1
    # via beautifulsoup4
stack-data==0.6.2
    # via ipython
terminado==0.17.1
    # via
    #   jupyter-server
    #   jupyter-server-terminals
    #   nbclassic
    #   notebook
tinycss2==1.2.1
    # via nbconvert
tornado==6.3.1
    # via
    #   ipykernel
    #   jupyter-client
    #   jupyter-server
    #   nbclassic
    #   notebook
    #   terminado
traitlets==5.9.0
    # via
    #   comm
    #   ipykernel
    #   ipython
    #   jupyter-client
    #   jupyter-core
    #   jupyter-events
    #   jupyter-server
    #   matplotlib-inline
    #   nbclassic
    #   nbclient
    #   nbconvert
    #   nbformat
    #   notebook
uri-template==1.2.0
    # via jsonschema
wcwidth==0.2.6
    # via prompt-toolkit
webcolors==1.13
    # via jsonschema
webencodings==0.5.1
    # via
    #   bleach
    #   tinycss2
websocket-client==1.5.1
    # via jupyter-server
